requirements.txt
paddlepaddle-gpu==2.6.1
paddleocr==2.7.3
protobuf==3.20.3
numpy==1.24.3
opencv-python-headless==4.8.1.78
streamlit